pyproject.toml
[tool.poetry]
name = "docker-dist-nn"
version = "0.1.0"
description = ""
authors = ["Tollan <tollanberhanu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "2.6.0"
torchvision = "0.21.0"
numpy = "^2.2.3"
docker = "^7.1.0"
tqdm = "^4.67.1"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
